fix: resolve Verus type coercion and mode errors
- Fix int/u64 division coercion in rate_limiter and strategies specs
- Replace saturating_mul with manual overflow checks in raft specs
- Inline exec function calls in ensures clauses to resolve mode errors
- Add uninterp keyword to generate_receipt_handle function
- Split compound ensures clauses using field accessor syntax

Resolves compilation errors allowing aspen-core to verify completely.

+++ crates/aspen-coordination/verus/rate_limiter_ops_spec.rs
@@ -533,7 +533,7 @@ verus! {
         requires refill_interval_ms > 0
         ensures
             current_time_ms >= last_refill_ms ==>
-                result == (current_time_ms - last_refill_ms) / refill_interval_ms,
+                result == ((current_time_ms - last_refill_ms) as int / refill_interval_ms as int) as u64,
             current_time_ms < last_refill_ms ==> result == 0
     {
         if current_time_ms >= last_refill_ms {
@@ -641,7 +641,7 @@ verus! {
         refill_interval_ms: u64,
     ) -> (result: u64)
         requires refill_interval_ms > 0
-        ensures result == (refill_amount * 1000) / refill_interval_ms ||
+        ensures result == ((refill_amount as int * 1000) / refill_interval_ms as int) as u64 ||
                 result == u64::MAX  // overflow case
     {
         // Use saturating multiplication to prevent overflow
